Factored application to Typescript
