+++ package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/react": "^18.0.12",
+    "@types/react-router-dom": "^5.3.3",
     "@unform/core": "^2.1.6",
     "@unform/web": "^2.1.6",
     "axios": "^0.21.1",
@@ -16,6 +18,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.2.1",
+    "typescript": "^4.7.3",
     "yup": "^0.32.9"
   },
   "scripts": {
@@ -44,6 +47,9 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^18.0.5",
+    "@types/react-modal": "^3.13.1",
+    "@types/styled-components": "^5.1.25",
     "json-server": "^0.16.3"
   }
 }
