+++ server.json
@@ -23,6 +23,14 @@
       "price": "25.90",
       "available": false,
       "image": "https://storage.googleapis.com/golden-wind/bootcamp-gostack/desafio-food/food3.png"
+    },
+    {
+      "image": "",
+      "name": "Caqui",
+      "price": "19.9",
+      "description": "",
+      "available": true,
+      "id": 4
     }
   ]
 }
