+++ src/App.tsx
@@ -4,13 +4,14 @@ import Routes from './routes';
 
 import GlobalStyle from './styles/global';
 
-const App = () => (
-  <>
+export function App(){
+  return(<>
     <GlobalStyle />
     <Router>
       <Routes />
     </Router>
   </>
-);
+  )
+};
 
 export default App;
