+++ src/components/Food/index.tsx
@@ -1,0 +1,96 @@
+import { FiEdit3, FiTrash } from 'react-icons/fi';
+
+import { Container } from './styles';
+import api from '../../services/api';
+import { useState } from 'react';
+
+interface Food{
+  id: number,
+  name: string,
+  description: string,
+  price: string,
+  available: boolean,
+  image: string
+}
+
+interface FoodProps{
+  food: Food,
+  handleEditFood: (food: Food)=>void,
+  handleDelete: (id: number)=>void
+}
+
+export function Food({ food, handleEditFood, handleDelete }: FoodProps){
+  const [isAvailable, setIsAvailable] = useState(food.available)
+  // constructor(props) {
+  //   super(props);
+
+  //   const { available } = this.props.food;
+  //   this.state = {
+  //     isAvailable: available
+  //   };
+  // }
+
+  async function toggleAvailable() {
+    await api.put(`/foods/${food.id}`, {
+      ...food,
+      available: !isAvailable,
+    });
+
+    setIsAvailable((oldIsAvailable)=> !oldIsAvailable );
+  }
+
+  function setEditingFood (){
+    handleEditFood(food);
+  }
+
+  return (
+    <Container available={isAvailable}>
+      <header>
+        <img src={food.image} alt={food.name} />
+      </header>
+      <section className="body">
+        <h2>{food.name}</h2>
+        <p>{food.description}</p>
+        <p className="price">
+          R$ <b>{food.price}</b>
+        </p>
+      </section>
+      <section className="footer">
+        <div className="icon-container">
+          <button
+            type="button"
+            className="icon"
+            onClick={setEditingFood}
+            data-testid={`edit-food-${food.id}`}
+          >
+            <FiEdit3 size={20} />
+          </button>
+
+          <button
+            type="button"
+            className="icon"
+            onClick={() => handleDelete(food.id)}
+            data-testid={`remove-food-${food.id}`}
+          >
+            <FiTrash size={20} />
+          </button>
+        </div>
+
+        <div className="availability-container">
+          <p>{isAvailable ? 'Disponível' : 'Indisponível'}</p>
+
+          <label htmlFor={`available-switch-${food.id}`} className="switch">
+            <input
+              id={`available-switch-${food.id}`}
+              type="checkbox"
+              checked={isAvailable}
+              onChange={toggleAvailable}
+              data-testid={`change-status-food-${food.id}`}
+            />
+            <span className="slider" />
+          </label>
+        </div>
+      </section>
+    </Container>
+  );
+};
