+++ src/components/Input/styles.ts
@@ -1,6 +1,13 @@
+import React from 'react';
 import styled, { css } from 'styled-components';
 
-export const Container = styled.div`
+
+interface Props{
+  isFocused: boolean,
+  isFilled: boolean
+}
+
+export const Container = styled.div<Props>`
   display: flex;
   align-items: center;
